Add fontification from `chatgpt-shell`
Add fontification logic from `chatgpt-shell`. This code should probably
live in `shell-maker` as also suggested by the `TODO` comments in
`chatgpt-shell`. However, for the time being it isn't and so we extract
the code into it's own library file. This allows us to easily remove it
once `shell-maker` is refactored.