Switch system prompt on-the-fly
While the system prompt can be switched via `customize`, it is simpler
to switch on-the-fly via `claude-shell-swap-system-prompt`. This is
similar to `chatgpt-shell-swap-system-prompt`.

In addition, customize will not immediately invalidate and re-render the
shell prompt.

In addition set function field of `claude-shell-system-prompt` to
retrieve the claude prompt from `claude-shell-system-prompts`. The
`:get` retrieval function in `defcustom` was very flaky.