Handle empty assistant responses
There can be a user input but an empty response, or not even a user
input. This e.g. happens when `shell-maker-interrupt` is called to
refresh the prompt.

When there is no response for an interaction the `cdr` of the history
alist as returned by shell-maker is `nil`. This is problematic for
constructing the request payload which expects a full user-assistant
interaction for the history.

Do not include history items if they only have a user prompt without an
assistant response, i.e. a history item with only a `car` but no `cdr`.